feat: generate the auth headers whenever the config is loaded, as opposed to only when the CLI is invoked
