+++ src/fides/core/user.py
@@ -1,17 +1,19 @@
 """Module for interaction with User endpoints/commands."""
 import json
-from os import getenv
-from os.path import isfile
-from pathlib import Path
 from typing import Dict, List, Tuple
 
 import requests
-import toml
-from pydantic import BaseModel
 
 from fides.cli.utils import handle_cli_response
 from fides.core.config import get_config
-from fides.core.utils import echo_green, echo_red
+from fides.core.utils import (
+    Credentials,
+    echo_green,
+    get_auth_header,
+    get_credentials_path,
+    read_credentials_file,
+    write_credentials_file,
+)
 from fides.lib.cryptography.cryptographic_util import str_to_b64_str
 from fides.lib.oauth.scopes import SCOPES
 
@@ -19,26 +21,6 @@ config = get_config()
 CREATE_USER_PATH = "/api/v1/user"
 LOGIN_PATH = "/api/v1/login"
 USER_PERMISSIONS_PATH = "/api/v1/user/{}/permission"
-
-
-class Credentials(BaseModel):
-    """
-    User credentials for the CLI.
-    """
-
-    username: str
-    password: str
-    user_id: str
-    access_token: str
-
-
-def get_credentials_path() -> str:
-    """
-    Returns the default credentials path or the path set as an environment variable.
-    """
-    default_credentials_file_path = f"{str(Path.home())}/.fides_credentials"
-    credentials_path = getenv("FIDES_CREDENTIALS_PATH", default_credentials_file_path)
-    return credentials_path
 
 
 def get_access_token(username: str, password: str, server_url: str) -> Tuple[str, str]:
@@ -55,49 +37,6 @@ def get_access_token(username: str, password: str, server_url: str) -> Tuple[str
     user_id: str = response.json()["user_data"]["id"]
     access_token: str = response.json()["token_data"]["access_token"]
     return (user_id, access_token)
-
-
-def write_credentials_file(credentials: Credentials, credentials_path: str) -> str:
-    """Write the user credentials file."""
-    with open(credentials_path, "w", encoding="utf-8") as credentials_file:
-        credentials_file.write(toml.dumps(credentials.dict()))
-    return credentials_path
-
-
-def read_credentials_file(
-    credentials_path: str,
-) -> Credentials:
-    """Read and return the credentials file."""
-    if not isfile(credentials_path):
-        raise FileNotFoundError
-    with open(credentials_path, "r", encoding="utf-8") as credentials_file:
-        credentials = Credentials.parse_obj(toml.load(credentials_file))
-    return credentials
-
-
-def create_auth_header(access_token: str) -> Dict[str, str]:
-    """Given an access token, create an auth header."""
-    auth_header = {
-        "Authorization": f"Bearer {access_token}",
-    }
-    return auth_header
-
-
-def get_auth_header(verbose: bool = True) -> Dict[str, str]:
-    """
-    Executes all of the logic required to form a valid auth header.
-    """
-    credentials_path = get_credentials_path()
-    try:
-        credentials = read_credentials_file(credentials_path=credentials_path)
-    except FileNotFoundError:
-        if verbose:
-            echo_red("No credentials file found.")
-        raise SystemExit(1)
-
-    access_token = credentials.access_token
-    auth_header = create_auth_header(access_token)
-    return auth_header
 
 
 def create_user(
